Spike a findMatchingLines async generator
This reads pages backwards from the end of the file, emitting the
pages for now, intending to emit only the matching lines.

Seems like it should work and keep the runtime memory small, though
the function will be large and fiddly, not sure the best way to
break it apart or even if that would be useful.