Implement findLatestLines
It's in the same ballpark as grep on large.log, even when there are
no matches and it has to consume the whole file. Not too shabby for
a javascript runtime.

I'm using 1M pages and allocating copy buffers for each line. Probably
not copying at that point and messing around with the page size could
optimize performance, but this seems reasonable.